+++ packages/components/autocomplete/src/autocomplete-item.tsx
@@ -49,7 +49,7 @@ export const AutocompleteItem = <T extends React.ElementType = "div">(
       className={cn(
         "relative flex w-full cursor-pointer select-none items-center gap-2 rounded-xl p-2 text-sm hover:bg-content2",
         className,
-        { "bg-action-hover": isActive },
+        { "bg-content2": isActive },
       )}
       {...restProps}
       {...getItemProps({