+++ packages/components/select/src/select-item.tsx
@@ -69,7 +69,7 @@ export const SelectItem = <T extends React.ElementType = "button">(
       className={cn(
         "relative flex w-full cursor-pointer select-none items-center gap-2 rounded-xl p-2 text-sm hover:bg-content2 focus-visible:ring-2 focus-visible:ring-primary",
         className,
-        { "bg-action-hover": isActive },
+        { "bg-content2": isActive },
       )}
       {...restProps}
       {...getItemProps({