+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ui/storybook
 
+## 0.0.135
+
+### Patch Changes
+
+- theme change
+- Updated dependencies
+  - @jamsr-ui/react@0.0.133
+
+## 0.0.134
+
+### Patch Changes
+
+- @jamsr-ui/react@0.0.132
+
 ## 0.0.133
 
 ### Patch Changes