+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/theme
 
+## 0.0.49
+
+### Patch Changes
+
+- theme change
+
+## 0.0.48
+
+### Patch Changes
+
+- added theme change
+
 ## 0.0.47
 
 ### Patch Changes